fix(l1): address review feedback — clippy, metrics pattern, monitor fixes
- fix clippy redundant closure in formatter.rs
- fix unnecessary u64 casts in sync_manager.rs
- migrate MetricsSync to default prometheus registry (Pattern B) so new
  sync metrics register once at init instead of per-gather. Removes
  gather_metrics() — exported via gather_default_metrics() automatically
- fix admin_syncStatus reporting stale phase after recoverable error
- fix on_failure() poll bypass — add force param so final poll executes
  even when instance status is already "failed"
- fix peer_top.sh Python 3.12+ f-string syntax — use str.format()
- remove dead DEGRADATION_STALL_TIMEOUT constant

+++ crates/networking/p2p/sync_manager.rs
@@ -141,11 +141,11 @@ impl SyncManager {
         }
 
         // Populate live progress from METRICS atomics
-        let headers = METRICS.downloaded_headers.get() as u64;
+        let headers = METRICS.downloaded_headers.get();
         let accounts_downloaded = METRICS.downloaded_account_tries.load(Relaxed);
         let accounts_inserted = METRICS.account_tries_inserted.load(Relaxed);
-        let storage_downloaded = METRICS.storage_leaves_downloaded.get() as u64;
-        let storage_inserted = METRICS.storage_leaves_inserted.get() as u64;
+        let storage_downloaded = METRICS.storage_leaves_downloaded.get();
+        let storage_inserted = METRICS.storage_leaves_inserted.get();
 
         if headers > 0 {
             diag.phase_progress
